# openapi.yaml
openapi: 3.0.3
info:
  title: NOC Chamados API
  description: |
    API REST para sistema de gestão de chamados de NOC (Network Operations Center).
    
    ## Autenticação
    A API usa autenticação JWT via HTTPOnly Cookie. Faça login primeiro usando `/auth/login`.
    
    ## Rate Limiting
    - Login: 5 tentativas por 15 minutos (por IP)
    
    ## Recursos
    - [Documentação Interativa](/api-docs)
    - [Guia Swagger](./SWAGGER_GUIDE.md)
    - [Coleção Postman](./NOC_Chamados.postman_collection.json)
    
  version: 1.2.3
  contact:
    name: Suporte NOC Chamados
  license:
    name: Proprietário
servers:
  - url: https://chamados.xmov.com.br/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento Local
...
paths:
...
  /chamados:
    get:
      tags:
        - Chamados
      summary: Listar Chamados
      description: Lista todos os chamados com filtros opcionais
      operationId: listarChamados
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por status
          schema:
            type: string
            enum: [aberto, em_atendimento, aguardando_operadora, fechado, todos]
            default: todos
        - name: impacto
          in: query
          description: Filtrar por impacto
          schema:
            type: string
            enum: [critico, alto, medio, baixo, todos]
            default: todos
        - name: operadoraId
          in: query
          description: Filtrar por operadora
          schema:
            type: integer
            example: 1
        - name: popId
          in: query
          description: Filtrar por POP
          schema:
            type: integer
            example: 1
        - name: busca
          in: query
          description: Termo de busca (multi-campo com AND)
          schema:
            type: string
            example: link fora
      responses:
        '200':
          description: Lista de chamados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chamado'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Chamados
      summary: Criar Chamado
      description: Cria um novo chamado
      operationId: criarChamado
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CriarChamadoLink'
                - $ref: '#/components/schemas/CriarChamadoTransporte'
      responses:
        '201':
          description: Chamado criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chamado'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chamados/{id}:
    get:
      tags:
        - Chamados
      summary: Obter Chamado
      description: Retorna detalhes de um chamado específico
      operationId: obterChamado
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do chamado
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Detalhes do chamado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChamadoDetalhado'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Chamados
      summary: Atualizar Chamado
      description: Atualiza informações de um chamado
      operationId: atualizarChamado
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarChamado'
      responses:
        '200':
          description: Chamado atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chamado'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Chamados
      summary: Deletar Chamado
      description: Exclui permanentemente um chamado (apenas supervisores)
      operationId: deletarChamado
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chamado excluído
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Chamado excluído com sucesso
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /chamados/{id}/periodos-impacto:
    get:
      tags:
        - Períodos de Impacto
      summary: Listar Períodos
      description: Lista períodos de impacto de um chamado
      operationId: listarPeriodosImpacto
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de períodos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PeriodoImpacto'

    post:
      tags:
        - Períodos de Impacto
      summary: Criar Período
      description: Adiciona um período de impacto ao chamado
      operationId: criarPeriodoImpacto
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarPeriodoImpacto'
      responses:
        '201':
          description: Período criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodoImpacto'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token armazenado em HTTPOnly cookie
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Não autorizado
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Acesso negado
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Dados inválidos
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso não encontrado
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
...
    Chamado:
      type: object
      properties:
        id:
          type: integer
          example: 1
        numero:
          type: string
          example: '2025.12345'
        status:
          type: string
          enum: [aberto, em_atendimento, aguardando_operadora, fechado]
          example: aberto
        impacto:
          type: string
          enum: [critico, alto, medio, baixo]
          example: critico
        tipoChamado:
          type: string
          enum: [link, transporte]
          example: link
        descricaoProblema:
          type: string
          example: Link fora do ar
        protocoloOperadora:
          type: string
          nullable: true
          example: INC123456
        dataDeteccao:
          type: string
          format: date-time
          example: '2025-12-15T10:00:00.000Z'
        dataAbertura:
          type: string
          format: date-time
          example: '2025-12-15T10:05:00.000Z'
        dataNormalizacao:
          type: string
          format: date-time
          nullable: true
          example: '2025-12-15T14:30:00.000Z'
        dataFechamento:
          type: string
          format: date-time
          nullable: true
        link:
          $ref: '#/components/schemas/Link'
          nullable: true
        transporte:
          $ref: '#/components/schemas/Transporte'
          nullable: true
        tipoFalha:
          $ref: '#/components/schemas/TipoFalha'
        abertoPor:
          $ref: '#/components/schemas/UsuarioSimples'
        fechadoPor:
          $ref: '#/components/schemas/UsuarioSimples'
          nullable: true
    ChamadoDetalhado:
      allOf:
        - $ref: '#/components/schemas/Chamado'
        - type: object
          properties:
            periodosImpacto:
              type: array
              items:
                $ref: '#/components/schemas/PeriodoImpacto'
    CriarChamadoLink:
      type: object
      required:
        - tipoChamado
        - linkId
        - tipoFalhaId
        - dataDeteccao
        - impacto
        - descricaoProblema
      properties:
        tipoChamado:
          type: string
          enum: [link]
        linkId:
          type: integer
          example: 1
        tipoFalhaId:
          type: integer
          example: 1
        dataDeteccao:
          type: string
          format: date-time
          example: '2025-12-15T10:00:00.000Z'
        impacto:
          type: string
          enum: [critico, alto, medio, baixo]
          example: critico
        descricaoProblema:
          type: string
          example: Link fora do ar
        protocoloOperadora:
          type: string
          example: INC123456
    CriarChamadoTransporte:
      type: object
      required:
        - tipoChamado
        - transporteId
        - tipoFalhaId
        - dataDeteccao
        - impacto
        - descricaoProblema
      properties:
        tipoChamado:
          type: string
          enum: [transporte]
        transporteId:
          type: integer
          example: 1
        tipoFalhaId:
          type: integer
          example: 1
        dataDeteccao:
          type: string
          format: date-time
        impacto:
          type: string
          enum: [critico, alto, medio, baixo]
        descricaoProblema:
          type: string
        protocoloOperadora:
          type: string
    AtualizarChamado:
      type: object
      properties:
        status:
          type: string
          enum: [aberto, em_atendimento, aguardando_operadora, fechado]
        protocoloOperadora:
          type: string
        descricao:
          type: string
        dataDeteccao:
          type: string
          format: date-time
        dataNormalizacao:
          type: string
          format: date-time
...
    PeriodoImpacto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        inicio:
          type: string
          format: date-time
          example: '2025-12-15T12:00:00.000Z'
        fim:
          type: string
          format: date-time
          nullable: true
          example: '2025-12-15T14:00:00.000Z'
        descricao:
          type: string
          nullable: true
          example: Horário de pico
        criadoPor:
          $ref: '#/components/schemas/UsuarioSimples'
    CriarPeriodoImpacto:
      type: object
      required:
        - inicio
      properties:
        inicio:
          type: string
          format: date-time
          example: '2025-12-15T12:00:00.000Z'
        fim:
          type: string
          format: date-time
          example: '2025-12-15T14:00:00.000Z'
        descricao:
          type: string
          example: Horário de pico
...
    Link:
      type: object
      properties:
        id:
          type: integer
          example: 1
        designador:
          type: string
          example: LINK-001
        tipoServico:
          type: string
          enum: [LINK IP, TRANSPORTE]
          example: LINK IP
        capacidade:
          type: string
          example: 10G
        diaVencimento:
          type: integer
          nullable: true
          minimum: 1
          maximum: 31
          example: 25
        ativo:
          type: boolean
          example: true
        enderecoPontaA:
          type: string
          nullable: true
        enderecoPontaB:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        operadora:
          $ref: '#/components/schemas/Operadora'
        pop:
          $ref: '#/components/schemas/Pop'
        transporte:
          $ref: '#/components/schemas/Transporte'
          nullable: true
...
    Transporte:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: TRANSPORTE-001
        fornecedor:
          type: string
          nullable: true
          example: Fornecedor A
        origem:
          type: string
          nullable: true
          example: São Paulo
        destino:
          type: string
          nullable: true
          example: Rio de Janeiro
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
        _count:
          type: object
          properties:
            links:
              type: integer
            chamados:
              type: integer
...
    Operadora:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Operadora A
        telefoneSuporte:
          type: string
          nullable: true
          example: '+55 11 1234-5678'
        emailSuporte:
          type: string
          format: email
          nullable: true
          example: suporte@operadora.com
        portalUrl:
          type: string
          format: uri
          nullable: true
          example: https://portal.operadora.com
        slaHoras:
          type: integer
          nullable: true
          example: 4
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
...
    Pop:
      type: object
      properties:
        id:
          type: integer
          example: 1
        codigo:
          type: string
          example: SP01
        nome:
          type: string
          nullable: true
          example: POP São Paulo
        cidade:
          type: string
          example: São Paulo
        estado:
          type: string
          minLength: 2
          maxLength: 2
          example: SP
        endereco:
          type: string
          nullable: true
          example: 'Rua Exemplo, 123'
        ativo:
          type: boolean
          example: true
...
    UsuarioSimples:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: João Silva
...
    TipoFalha:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Fibra Rompida
        descricao:
          type: string
          nullable: true
          example: Cabo de fibra óptica rompido
...
security:
  - cookieAuth: []